move study save and load to io module
